refactor: 코드 전역에서 사용되는 기능은 common 패키지로 통합
