+++ src/main/java/courseitda/auth/ui/interceptor/AuthRoleCheckInterceptor.java
@@ -4,8 +4,8 @@ import courseitda.auth.domain.AuthRole;
 import courseitda.auth.domain.AuthTokenExtractor;
 import courseitda.auth.domain.AuthTokenProvider;
 import courseitda.auth.domain.RequiresRole;
-import courseitda.exception.auth.AuthenticationException;
-import courseitda.exception.auth.AuthorizationException;
+import courseitda.common.exception.auth.AuthenticationException;
+import courseitda.common.exception.auth.AuthorizationException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Arrays;
