+++ src/main/java/courseitda/common/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package courseitda.config;
+package courseitda.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
