+++ src/main/java/courseitda/common/exception/BadRequestException.java
@@ -1,7 +1,7 @@
-package courseitda.exception;
+package courseitda.common.exception;
 
 public class BadRequestException extends RuntimeException {
-    public BadRequestException(String message) {
+    public BadRequestException(final String message) {
         super(message);
     }
 }
