+++ src/main/java/courseitda/common/exception/BusinessRuleException.java
@@ -1,7 +1,7 @@
-package courseitda.exception;
+package courseitda.common.exception;
 
 public class BusinessRuleException extends RuntimeException {
-    public BusinessRuleException(String message) {
+    public BusinessRuleException(final String message) {
         super(message);
     }
 }
