+++ src/main/java/courseitda/common/exception/NotFoundException.java
@@ -1,7 +1,7 @@
-package courseitda.exception;
+package courseitda.common.exception;
 
 public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
+    public NotFoundException(final String message) {
         super(message);
     }
 }
