+++ src/main/java/courseitda/common/exception/auth/AuthTokenNotFoundException.java
@@ -1,0 +1,8 @@
+package courseitda.common.exception.auth;
+
+public class AuthTokenNotFoundException extends RuntimeException {
+
+    public AuthTokenNotFoundException(final String message) {
+        super(message);
+    }
+}
