+++ src/main/java/courseitda/common/exception/auth/AuthorizationException.java
@@ -1,0 +1,8 @@
+package courseitda.common.exception.auth;
+
+public class AuthorizationException extends RuntimeException {
+
+    public AuthorizationException(final String message) {
+        super(message);
+    }
+}
