+++ src/main/java/courseitda/common/exception/resource/AlreadyExistException.java
@@ -1,0 +1,8 @@
+package courseitda.common.exception.resource;
+
+public class AlreadyExistException extends RuntimeException {
+
+    public AlreadyExistException(final String message) {
+        super(message);
+    }
+}
