+++ src/main/java/courseitda/common/exception/resource/ResourceInUseException.java
@@ -1,0 +1,8 @@
+package courseitda.common.exception.resource;
+
+public class ResourceInUseException extends RuntimeException {
+
+    public ResourceInUseException(final String message) {
+        super(message);
+    }
+}
