--- src/main/java/courseitda/exception/auth/AuthTokenNotFoundException.java
@@ -1,8 +1,0 @@
-package courseitda.exception.auth;
-
-public class AuthTokenNotFoundException extends RuntimeException {
-
-    public AuthTokenNotFoundException(String message) {
-        super(message);
-    }
-}
