--- src/main/java/courseitda/exception/auth/AuthorizationException.java
@@ -1,8 +1,0 @@
-package courseitda.exception.auth;
-
-public class AuthorizationException extends RuntimeException {
-
-    public AuthorizationException(String message) {
-        super(message);
-    }
-}
