--- src/main/java/courseitda/exception/resource/AlreadyExistException.java
@@ -1,8 +1,0 @@
-package courseitda.exception.resource;
-
-public class AlreadyExistException extends RuntimeException {
-
-    public AlreadyExistException(String message) {
-        super(message);
-    }
-}
