--- src/main/java/courseitda/exception/resource/ResourceInUseException.java
@@ -1,8 +1,0 @@
-package courseitda.exception.resource;
-
-public class ResourceInUseException extends RuntimeException {
-
-    public ResourceInUseException(String message) {
-        super(message);
-    }
-}
