+++ src/main/java/courseitda/placesearch/domain/SearchedPlace.java
@@ -1,7 +1,7 @@
 package courseitda.placesearch.domain;
 
-import courseitda.exception.BusinessException;
-import courseitda.exception.ErrorCode;
+import courseitda.common.exception.BusinessException;
+import courseitda.common.exception.ErrorCode;
 
 public record SearchedPlace(
         String name,
