+++ src/main/java/courseitda/workspace/application/RepresentativeCategoryPlaceService.java
@@ -1,7 +1,7 @@
 package courseitda.workspace.application;
 
 import courseitda.auth.domain.MemberAuthInfo;
-import courseitda.exception.NotFoundException;
+import courseitda.common.exception.NotFoundException;
 import courseitda.workspace.domain.Category;
 import courseitda.workspace.domain.CategoryPlace;
 import courseitda.workspace.domain.CategoryPlaceRepository;
